Keep items somebody has favorited locally
git-svn-id: https://svn.midgard-project.org/midgard/trunk/midcom@11585 62f49d1c-1c16-0410-9156-d311ce72abac

+++ net.nemein.rss/fetch.php
@@ -331,6 +331,17 @@ class net_nemein_rss_fetch extends midcom_baseclasses_components_purecode
             if (!in_array($item->$guid_property, $item_guids))
             {
                 // This item has been removed from the feed.
+                
+                // If it has been favorited keep it
+                $_MIDCOM->componentloader->load_graceful('net.nemein.favourites');
+                $qb = net_nemein_favourites_favourite_dba::new_query_builder();
+                $qb->add_constraint('objectGuid', '=', $item->guid);
+                if ($qb->count_unchecked() > 0)
+                {
+                    continue;
+                    // Skip deleting this one
+                }
+                
                 $item->delete();
             }
         }
